requirements.txt
flask
flask-sqlalchemy
flask_wtf
flask-migrate
python-dotenv
rav
